put sequelize store in config
