+++ config/connection.js
@@ -1,6 +1,8 @@
 require('dotenv').config();
 
 const Sequelize = require('sequelize');
+const session = require('express-session');
+const SequelizeStore = require('connect-session-sequelize')(session.Store);
 
 const sequelize = process.env.JAWSDB_URL
     ? new Sequelize(process.env.JAWSDB_URL)
@@ -10,4 +12,18 @@ const sequelize = process.env.JAWSDB_URL
         port: 3306,
         
     });
-    module.exports = sequelize;
+    //cookies 
+const sess = {
+    secret: process.env.SESSION_SECRET,
+    resave:false,
+    saveuninitialized:true,
+    cookie:{},
+    store: new SequelizeStore({
+        db:sequelize
+    })
+
+};
+    module.exports = {
+        sequelize,
+        sess
+    };
